Clean up CSS loaders
