--- plugins/lfa-compilation/web_modules/css-minify-loader.js
@@ -1,9 +1,0 @@
-var CleanCSS = require('clean-css');
-
-module.exports = function (content) {
-  this.cacheable(true);
-  var opts = {
-    restructuring: false
-  };
-  return new CleanCSS(opts).minify(content).styles;
-};
